Fix ClusterInfo type ResourceExport recreation bug
After Gateway HA is enabled, the ClusterInfo type of ResourceExport
will be recreated when the active Gateway is changed. But there is a case
that a new ClusterInfo of ResourceExport creation may fail when the
leader controller process is slow and existing ResourceExport is not
deleted in time.
Fix the issue through retry when the existing ResourceExport's DeletionTimestamp
is not zero.

Signed-off-by: Lan Luo <luola@vmware.com>