Adds the containerize option for antrea-agent installation on Linux VMs
The installation script would help to install antrea-agent service
in containers by specifying the container option in the installation
command for Linux VMs.

Install Docker automatically based on the Linux OS.
The script picks the latest Antrea version or the user can
also specify the version.
This creates systemd service which monitors the containers.
Also support start, restart and stop the service.

Signed-off-by: nithishs <nithishs@vmware.com>