Fix typo in S3Uploader (#4433)
This commits fixes a typo in S3Uploader: when an upload error occurs,
we should discard the successfully-uploaded buffers and keep the
rest in `buffersToUpload`. Previously we were mistakenly modifying
`bufferQueue`, instead of `buffersToUpload`. The typo was not caught
by UT, as we currently do not have a test case to properly simulate a
partial-success upload.

Signed-off-by: heanlan <hanlan@vmware.com>

Signed-off-by: heanlan <hanlan@vmware.com>