Make vm agent CI script install antrea-agent service.
The CI currently uses a customized image, which contains the antrea-agent
service pre-created. The CI/Test only starts the antrea-agent service and
validates the status of the antrea agent service. But this commit integrates
the installations script with CI, utilizes that installation script
(like an end user), and installs the antrea-agent on both Linux and Windows VMs
and CI validates the status of the antrea-agent service in both VMs.

Also, Now the CI would support installing antrea-agent using install script
on multiple Linux and Windows VMs for testing purposes.

Signed-off-by: nithishs <nithishs@vmware.com>