Optimize the updating of Endpoints for Service in AntreaProxy
Before this PR, when updating the Endpoints of a Service in AntreaProxy,
the corresponding OVS group for the Service will be reinstalled completely,
which means every bucket which is for an Endpoint in the group will be
reinstalled regardless of whether it is a new Endpoint or not. In this PR,
the Endpoints to be added or removed will be synced by OVS group operations
INSERT_BUCKET or REMOVE_BUCKET introduced in Openflow 1.5, instead of
reinstalling the whole corresponding OVS group.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>