Change MCNP secFlows implementation
This PR changes how we implement MultiClusterNetworkPolicy security
flows.

Before, we created a security rule for each MultiClusterNetworkPolicy
rule to install security flows.

Now, we use the default drop flow to implement the security flows. For
a MultiClusterNetworkPolicy rule, the drop flow will use the
UnknownLabelIdentity and the original destination matching to drop
packets.

Signed-off-by: graysonwu <wgrayson@vmware.com>