Stop using ClusterFirstWithHostNet DNSPolicy for antrea-agent (#4548)
The DaemonSet for a CNI should not use ClusterFirstWithHostNet and
should not have a dependency on cluster DNS, since cluster DNS itself
depends on the CNI. Therefore, we revert to the "default" DNSPolicy for
host-network Pods.

The only dependency we had on cluster DNS was in the Flow Exporter,
where we would try to connect to the Flow Aggregator using DNS name
"flow-aggregator.flow-aggregator.svc" (by default). This would only work
on Linux, and not on Windows (for Windows Nodes, the Cluster IP had to
be provided instead). We update the Flow Exporter to resolve the Service
Cluster IP using the K8s API instead.

Notable changes:
* go-ipfix is upgraded to v0.6.0 to support configuring the server name
  for Flow Aggregator certificate verification in the Flow Exporter
* the flowAggregatorAddress parameter is no longer required in the Flow
  Aggregator configuration
* the format for the flowCollectorAddr parameter in the Flow Exporter
  configuration (antrea-agent) is changed in a non backwards-compatible
  way. The "host" name must be provided as
  "flow-aggregator/flow-aggregator" instead of
  "flow-aggregator.flow-aggregator.svc".

Fixes #3279

Signed-off-by: Antonin Bas <abas@vmware.com>