Replace custom Clair scan with Trivy Github Workflows (#4541)
We have had ci/clair-scan for a while, but I don't think it has proven
useful as we scarcely look at the results, and vulnerabilities are
usually reported by Antrea users and third-parties. The way we run Clair
is also probably outdated, as this code has not been updated recently.

Since we recently added a Makefile target to run Trivy scans, we replace
our "custom" Clair scan code, as well as our Github workflow running
scans, with new Github workflows using the standard
aquasecurity/trivy-action Github action.

We have 2 Github worfklows running Trivy on the Antrea Ubuntu-based
image. The first one runs daily (or on demand) and generates reports
(which are then uploaded as artifacts) for the latest Antrea image as
well as the image for the latest Antrea released version. The second one
runs whenever a Pull Request is created to release a new Antrea version.

Signed-off-by: Antonin Bas <abas@vmware.com>