Update Egress API to support multiple Egress IPs and pools
Nodes in a cluster could reside in multiple subnets, but the egress IP
needs to be routable in the underlay network and usually resides in the
same subnet of the Node hosting it. Therefore, there may be a situation
where no available nodes are eligible to host an egress IP address if
all Nodes in a subnet are down, interrupting egress traffic of the
workloads that use the IP.

As the first step of supporting the above scenario, the patch extends
the Egress IP to support multiple Egress IPs and pools so that one
Egress IP can be configured for each subnet, making Egress failover
across the whole cluster possible.

Besides, it also adds a field, `status.egressIP`, to represent the
effective Egress IP. When there is no eligible Node for any of the
Egress IPs, the field will be empty.

Signed-off-by: Quan Tian <qtian@vmware.com>