Refine Endpoint selection for multi-cluster Service
When the Endpoint of Multi-cluster Service is a local Service ClusterIP,
refine the action to let it go to the corresponding exported Service's
group to do final Endpoint selection. This can avoid the case that the
traffic goes out of antrea-gw0 and goes back to OVS again when a
local Pod is trying to access a MC Service but a local Service's
Endpoint is selected.

Signed-off-by: Lan Luo <luola@vmware.com>