Use new Github Action to upload release assets
Use alexellis/upload-assets instead of actions/upload-release-asset,
which is no longer maintained. One advantage is that we can list all
assets with a glob expression, so the workflow file is much less
verbose.

We also do a minor change to ./hack/generate-manifest-windows.sh, so
that it can be run on macOS (one of the sed commands was GNU-specific,
and there is no reason to use sed instead of kustomize).

Fixes #4514

Signed-off-by: Antonin Bas <abas@vmware.com>