Fix e2e TestAntctl on windows testbed
TestAntctl/testAntctlControllerRemoteAccess failed
on all windows testbed because it cannot get
output of `antctl get memberlist` while collecting
supportbundle from outside on windows testbed as
memberlist is not running.

This PR handles the above case as success on linux
and writes output to file that memberlist is not running.

On Windows, don't collect memberlist as it never runs on Windows.

Fixes #4659

Signed-off-by: Kumar Atish <atish.iaf@gmail.com>