Add Controller configuration to specify client CA
In the existing logic, Antrea Controller retrieves the client CA from ConfigMap
kube-system/extension-apiserver-authentication, which is published by kube-apiserver.
In some cases, the incoming connection is from the client with a certificate signed
by a different CA bundle.

This patch adds a configuration in antrea-controller to allow user to specify the client
CA. Antrea Controller will use it to validate the incoming client certificate in a mTLS
connection.

Signed-off-by: wenyingd <wenyingd@vmware.com>