Add ServiceLocalToRemote reject type (#4547)
For single cluster rejection, there is no difference for Service traffic or Pod-to-Pod traffic in localToRemote rejection. Service LB won't be executed on the Node where we generate reject responses. So we didn't define podLocalToRemote and serviceLocalToRemote and just let the reject response packet start from L3/L2Forwarding table based on the Node type.

But in multi-cluster situations, when the server Endpoint Pod is on the gateway Node of its cluster, where Service LB is executed, we need reject response packets to go thru the whole pipeline to be correctly UnDNATed and forwarded.