Refine Endpoint selection for Multi-cluster Service (#4693)
* Revert "Refine Endpoint selection for multi-cluster Service (#4508)"

This reverts commit 0729d6a8b569cec4a59b95f48cc0d3408a10ec75.

Signed-off-by: Lan Luo <luola@vmware.com>

* Refine Endpoint selection for MC Service

Add a new flow for the Service's ClusterIP in the EndpointDNAT table with
group action. When an Endpoint of a Multi-cluster Service is a local Service
ClusterIP and being selected, it will go to the corresponding exported Service's
group to select the final Endpoint. This can avoid that the traffic goes out of the
OVS bridge from antrea-gw0 (and handled by kube-proxy when it is running) and
comes back again.

The proposal details can be found in the comment:
https://github.com/antrea-io/antrea/pull/4508#discussion_r1130308109

Signed-off-by: Lan Luo <luola@vmware.com>

---------

Signed-off-by: Lan Luo <luola@vmware.com>