Fix ClusterIP Service processing
AntreaProxy doesn't allocate a group for cluster Endpoints if a
ClusterIP Service's InternalTrafficPolicy is Local, so getting the group
ID of cluster Endpoints returns nothing. The previous code stopped
processing the Service and logged error mistakenly in this case, causing
this Service's flows and routes never being deleted.

The patch fixes it and enhances the validation for this case.

Signed-off-by: Quan Tian <qtian@vmware.com>