Avoid duplicate Node Results in Live Traceflow Status
After receiving a Packet In for a live Traceflow, the controller
uninstalls the corresponding OVS flows to prevent additional Packet Ins
(from different connections which also match the Live Traceflow
filters).

However, if 2 connections happen within a very short time window (< 1ms
in my testbed) and both match the Live Traceflow filters, it is still
possible for 2 Packet In messages to be received. To avoid duplicate
Node Results in the Live Traceflow Status, we need to ignore all Packet
In messages received after the first one.

Fixes #4714

Signed-off-by: Antonin Bas <abas@vmware.com>