Update codecov.yml (#4226)
1. Codecov do not comment until at least 5 builds have been uploaded
from the CI pipeline. By default codecov will post and/or update the
pull request comment after it processes each report uploaded for a
particular pull request commit. As the Antrea CI process uploads many
reports to codecov, this can be confusing for team members viewing the
pull request as reports are processing.
2. patch/status ensures the patch lines cov > 65%(ut or e2e/integration).
3. project/status ensures the project ut cov not decrease.
4. Set carryforward: false, as the Carryforward Flags are designed for
projects that do not upload total coverage for every commit.

Signed-off-by: Wenqi Qiu <wenqiq@vmware.com>