Support inspecting fragmented DNS message for FQDN policy (#4732)
A DNS message sent over TCP will be fragmented if its length is greater
than MTU, in which case FQDN policy wouldn't work as expected because
the used DNS library expects the entire message. However, it's been
observed that DNS message was fragmented frequently in tests, which may
be applicable to production environments as well. On the other hand, we
don't really need the entire message to support FQDN policy, and the
first fragment normally already contains the information we need, the
question and answer sections.

This patch forks and modifies the message Unpack function to be able to
handle fragmented messages.

Signed-off-by: Quan Tian <qtian@vmware.com>