Fallback to Endpoints API if EndpointSlice API is unavailable (#4741)
When EndpointSlice feature is enabled, it required EndpointSlice v1 API
to be available, otherwise it didn't work. But EndpointSlice v1 API is
only available in Kubernetes >= 1.21. For earlier versions, there is no
simple way to fallback to EndpointSlice v1beta1 API because the proxy
library code supports one API version only.

To make Antrea compatible with earlier releases, we could check if the
EndpointSlice v1 API is available and fallback to Endpoints API if it's
not.

Signed-off-by: Quan Tian <qtian@vmware.com>