Remove hack/netpol and associated Kind CI job
hack/netpol was always meant as a temporary placeholder, while new
Netpol tests were being contributed to upstream K8s. We have been
meaning to remove the folder for 2 years.

Note that the jenkins-networkpolicy CI job runs a recent version of K8s
e2e tests (1.25+), which includes the Netpol test suite. We could also
consider adding a new Kind CI job (as a Github workflow) dedicated to
running the upstream Netpol tests, but it is not strictly necessary and
could be an extra burden on our Github CI, as the tests take some time
to complete.

Fixes #1740

Signed-off-by: Antonin Bas <abas@vmware.com>