Stop serving v1alpha2 version of the ClusterGroup CRD
The API was scheduled for deletion a year ago.
My proposal is to stop serving the API in the next release (v1.12), and
remove it completely in the release after that (v1.13). If someone
upgrades to Antrea v1.12 and gets errors, they can easily edit the CRD
to re-enable serving for the older version (and then convert all their
existing resources).

Signed-off-by: Antonin Bas <abas@vmware.com>