Fix release badge at the top of README file (#4818)
It is now possible to display the latest release by semver, instead of
the latest release by date. When displaying by date, the README may show
an older Antrea version, for which a patch release was recently
published. Instead, we prefer to display the most recent patch release
for the most recent minor release, which is what most new users will
need / want to deploy.

Signed-off-by: Antonin Bas <abas@vmware.com>