Stop exporting KUBECONFIG in ci/test-conformance-gke.sh (#4873)
This is causing the cloud-antrea-gke-cleanup Jenkins job to fail (even
though the cluster is deleted properly), because for some reason "gcloud
container clusters delete" tries to create the Kubeconfig file (and the
directory does not exist):

ERROR: (gcloud.container.clusters.delete) Unable to create private file [/home/ubuntu/jenkins/out/gke/kubeconfig]: [Errno 1] Operation not permitted: '/home/ubuntu/jenkins/out/gke/kubeconfig'

This issue was introduced in #4862. It was not caught during testing
because the cleanup job, which is triggered by the main conformance job,
was not triggered using the correct parameters. Current build parameters
(ANTREA_REPO and ANTREA_GIT_REVISION) need to be propagated to the
cleanup job to enable proper testing before merges. We update the cloud
job definitions so that current-parameters is set to true when
triggering the cleanup jobs.

Signed-off-by: Antonin Bas <abas@vmware.com>