Add new exporter to FlowAggregator to write flows to local file (#4855)
Flows are written to a local "log" file, in CSV format, with support for
log rotation. Not all the fields from the flow records are
included. Configuration options include the ability to filter flows
based on ingress / egress network policy rule actions (in the future,
additional filtering capabilities could be introduced).

For #3794

Signed-off-by: Antonin Bas <abas@vmware.com>