Install Suricata from source in base Docker image
The latest release from Suricata could suffer from a bug, impacting
Antrea features. When installing from PPA, we don't have flexibility
when it comes to the Suricata version we ship with Antrea.

Instead, we can install Suricata from source as part of our Docker
build.

The advantages are:

* full control over the version we install.
* a smaller Antrea image, as we do not need to install all of Suricata's
  dependencies / enable all its features.

The disadvantages of building from source are:

* CVEs in Suricata itself won't be detected by scanners (CVEs in
  Suricata's dependencies will).
* While we have more customization options, we also have to manually
  keep track of Suricata's dependencies (and use correct build options).
* We have to remember to update the Suricata version we build & install.

For #4921

This can also be treated as a temporary fix until Suricata 6.0.12 is
released, at which point we could revert this change.

Signed-off-by: Antonin Bas <abas@vmware.com>