Fix GKE CI job
The gke-gcloud-auth-plugin binary is now required (in addition to the
main gcloud binary).

As a result, we make some changes to the test-conformance-gke.sh script.
The script now takes a '--gcloud-sdk-path' flag, which points to the
installation directory of the gcloud SDK. The bin directory from the SDK
is then added to the path.

When running the script locally, one can omit the '--gcloud-sdk-path'
flag, as long as the PATH already include the SDK bin directory (which
is the default hen following standard installation instructions).

Signed-off-by: Antonin Bas <abas@vmware.com>