Change to use PacketIn2/Controller2 (#4768)
This commit replaces the use of packetIn and controller with
packetIn2 and controller2, respectively. It also replaces the use of
the 'reason' with the 'userdata' to distribute packetIn2 to the
handler. Additionally, it replaces the use of 'customreason'
with 'userdata' to indicate what operations are needed for the
handler, when there is a chance that multiple operations are required
for a single packetIn2.

Furthermore, this commit introduces a change to use 'pause'+`resume`
for FQDN policy.

This commit upgrades ofnet to v0.7.1 and libOpenflow to v0.10.1, where
the packetIn2/controller2 support were added.

Signed-off-by: graysonwu <wgrayson@vmware.com>