Cache OpenFlow modification message in Antrea Agent (#4495)
The motivation for this patch is to reduce the memory used for storing
the installed OpenFlow flow entries to replay them when OVS is re-connected.
In the existing code, struct ofctrl.Flow is stored in the memory, which
has taken a lot of memory because all kinds of the candidate Match fields
are allocated in one such object although they may not use in the final Flow.

This patch is to store the FlowModification message as a substituation of the
ofctrl.Flow objects. As the FlowModification message is the obejct which
was sent to OVS via the OF connections in final, it contains enough information
to replay the OpenFlow flow entry. Besides, a FlowModification only contains
the necessary fields which are set with values.

A test with this patch may reduce about 140M memory on a setup with about
130K OpenFlow entries in antrea-agent (from 380M to 240M)

Signed-off-by: wenyingd <wenyingd@vmware.com>