Fix flaky TestUpdateEgress
The test is flaky (fails from time to time in Github CI) because the
Egress Controller relies on the CR's Generation metadata field to be
incremented when a change has been made to the Egress Spec. A specific
sequence of events in the TestUpdateEgress unit test can lead to a Spec
change being ignored by the Controller, causing the test to fail. In a
normal deployment, incrementing the Generation field is handled by the
K8s apiserver, but when using the fake client, this is not implemented
by design.

To avoid the issue, we add "reactors" for the update and patch actions,
which are in charge of incrementing the Generation field. Unfortunately,
this requires accessing the fake client's object tracker directly. Note
that we always increment the Generation field no matter what the actual
change is (while the K8s apiserver only increments it for actual Spec
changes). It should not matter for our unit tests.

Fixes #5027

Signed-off-by: Antonin Bas <abas@vmware.com>