Fix hack/make-metrics-doc.sh on macOS
It appears that the script was never really working on macOS, unless the
built-in fmt and sed utilities were replaced by their GNU
"equivalent". Instead of making this assumption, we explicitly use gfmt
and gsed when macOS is detected, and give an error if they have not been
installed.

While in other scripts, we try to use expressions which are compatible
with both flavors of sed (GNU and FreeBSD), this script is used so
rarely that it doesn't seem worth the maintenance effort. In particular,
there is no simple equivalent to the GNU sed expression to capitalize
every word in a sentence.

Fixes #5012

Signed-off-by: Antonin Bas <abas@vmware.com>