Use OpenFlow group for Network Policy logging
OpenFlow groups are created for Network Policy logging purpose, which are used
as the OpenFlow action if logging is enabled on the related resource.

Having tested with the worst case that enabling logging on an Antrea-native
policy rule to drop all UDP packets, which expects that all UDP packets are sent
to antrea-agent via packetIn channel, OVS CPU usage is changede from about 1.0%
to 1.3% comparing to the previous implementations in which packet is firstly
dropped by OVS meter before applying the policy actions and logging actions.
The CPU usage increasement introduced by this change is acceptable.

Besides, this change also explicitly set parameter enableLogging as "false" in
function "defaultDropFlow" when it is called by "flowsToTrace". This is to
avoid generating a unnecessary logging packet in a trace flow request.

Signed-off-by: wenyingd <wenyingd@vmware.com>