Fix flaky TestClean unit tests for storagebundle REST server
The test was using a Sleep statement, which was actually too short for
some CI machines, causing flakes.

Instead we use a virtual clock for the test, which means that the test
runtime is no longer dependent on the "duration" parameter for the
`clean` function. We also use polling to validate assertions.

Signed-off-by: Antonin Bas <abas@vmware.com>