Fix flaky e2e tests (#4989)
In e2e tests like TestProxyHairpinIPv4:

- A test Service is created
- A test Pod for the Service is created
- Probe the Service using code `data.RunCommandFromPod(data.testNamespace, pod, container, []string{"wget", "-O", "-", url, "-T", "5"})`

When using the command wget -O url -T 5 to connect to a Service without an Endpoint,
which is implemented by AntreaProxy, a rejection message will be received, and no retry
will be attempted. There is a chance that even if the test Pod is ready, the Service may
not be fully synced, resulting in a rejection message and a failed test. This PR fixes
the flaky test by adding retry logic.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>