Refine antctl mc deploy command
There is a bug that when user deploys Antrea Multi-cluster via `antctl mc`
on EKS clusters, user can't create ClusterSet and ClusterClaims with an
error like below:
```
Internal error occurred: failed calling webhook "vclusterclaim.kb.io":
failed to call webhook: ... x509: certificate signed by unknown authority
```
It's caused by a slow deployment when using `antctl mc deploy`,
antrea-mc-controller is running but validation webhook configuration is
not created yet, the `cabundle` in the webhook configuration won't be updated.
It turns out `GetAPIGroupResources` is a time-consuming step and it's
also unnecessary to be called every time in a loop, so move it out to the
caller with a few other refinements to fix the bug.

Signed-off-by: Lan Luo <luola@vmware.com>