Update short-names for AntreaNetworkPolicy as ANNP
As AdminNetworkPolicy will be supported and its short-name
is ANP, which collides with the current short-name for
AntreaNetworkPolicy, this solution updates short-names
for AntreaNetworkPolicy to ANNP. Meaning Antrea Namespace
Network Policy, it also corresponds with ACNP.

Existing short-name "anp" is deprecated for AntreaNP, and
will be removed in the next release.

Fixes #5063.

Signed-off-by: Qiyue Yao <yaoq@vmware.com>