Configure Pod MTU on Agent startup
Currently, the Pods' MTU will not be updated once created.
This will cause issues when the Pod's MTU need to be changed.
e.g. User redeploys Antrea with IPsec or WireGuard enabled, or
user sets the default MTU to a different value. For now, user
must restart all Pods to set the correct MTU. This patch will
fix this issue by setting the Pod's MTU on agent startup.

Signed-off-by: Xu Liu <xliu2@vmware.com>