Bugfix: Bump up libOpenflow version
This is to resolve an issue when unmarshaling IP packet which may lead to
incorrect field values in PacketIn meesages. The issue is because the existing
code uses a pointer reference to set byte slice fields when unmarshaling the
mesages, and libOpenlow reuses the buffer to process the next messages.

Signed-off-by: wenyingd <wenyingd@vmware.com>