Bump up libOpenflow and ofnet versions
This is to resolve an issue in packetIn2 which may lead to incorrect field
values in PackintIn2.Userdata and PacketIn2.Metadata. The issue is because the
existing code uses a pointer reference to set byte slice fields when
unmarshaling the mesages, and libOpenlow reuses the buffer to process the next
messages.

Signed-off-by: wenyingd <wenyingd@vmware.com>