Fix Egress IP not advertised in some cases (#5142)
This patch fixes two cases, in which an Egress IP is not advertised:

- When userspace ARP responder is not running (arp_ignore is 0),
  assigning a new Egress IP to a Node would not advertise the IP.
- When an Egress IP is obtained by multiple Nodes in some situations,
  e.g. split brain, the eventual winner would not advertise the IP after
  they are back to normal, because the IP is already assigned to the
  Node.

The patch moves IP advertising out of responders. Responders will only
be responsible for answering neighbor queries. The IP assigner will
advertise an IP when the IP is newly assigned to the Node, or the IP's
Node recorded in Egress API is updated from another Node to this Node.

Signed-off-by: Quan Tian <qtian@vmware.com>