DISABLE CGO while building binaries
The e2e test for antctl currently copy the antctl binary from the Docker
container to the host Node, and run the binary from the host. The host
may be missing some system libraries that are required to run the binary,
so we need to disable CGO to avoid such dependencies and ensure that the
binary is portable.

Signed-off-by: Pulkit Jain <jainpu@vmware.com>